feat: put tests in parent workspace
