--- falcon2/jest.config.js
@@ -1,6 +1,0 @@
-/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
-module.exports = {
-    preset: "ts-jest",
-    testEnvironment: "node",
-    modulePathIgnorePatterns: ["old"],
-};
