+++ falcon2/package.json
@@ -6,7 +6,6 @@
   "author": "cmudig",
   "license": "MIT",
   "scripts": {
-    "test": "jest",
     "build": "tsc",
     "format": "prettier --write src"
   },
@@ -29,9 +28,7 @@
     "@types/jest": "^29.2.1",
     "@types/ndarray": "1.0.11",
     "@types/ndarray-ops": "^1.2.4",
-    "jest": "^29.2.2",
     "prettier": "^2.7.1",
-    "ts-jest": "^29.0.3",
     "typescript": "^4.8.4"
   }
 }
