+++ jest.config.js
@@ -1,0 +1,10 @@
+/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
+module.exports = {
+	testEnvironment: "node",
+	preset: "ts-jest/presets/default-esm",
+	globals: {
+		"ts-jest": {
+			useESM: true,
+		},
+	},
+};
